requirements_donut.txt
accelerate==1.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
filelock==3.19.1
fsspec==2025.7.0
hf-xet==1.1.7
huggingface-hub==0.34.4
idna==3.10
jinja2==3.1.6
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.32.0
psutil==7.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.6.2
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tokenizers==0.21.4
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.55.2
triton==3.4.0
typing-extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
